Allow overriding diff binary name
On SmartOS the `diff` command doesn't understand the -N flag. Instead we would prefer to use the
`gdiff` binary, which does understand the -N flag (and is GNU diff, so behaves the same as capistrano
expects.)

This change now looks up the diff binary's name in a capistrano variable, which we default
to being `diff` so it's backwards compatible out the box. Add `set :faster_assets_diff, :gdiff`
to your deploy.rb to override the binary name.